refactor: rename invalid package declaration
